backend/requirements.txt
autopep8
attrs
backcall
certifi
chardet
decorator
Django
django-environ
django-extensions
flake8
flake8-commas
flake8-mypy
gevent
greenlet
uWSGI
idna
ipython
ipython-genutils
jedi
mccabe
mypy
parso
pexpect
pickleshare
Pillow
prompt-toolkit
psycopg2
ptyprocess
pycodestyle
pyflakes
Pygments
pytz
requests
rope
simplegeneric
six
traitlets
typed-ast
urllib3
wcwidth
Werkzeug
elasticsearch
django-cors-headers
django-allow-cidr
djangorestframework
django-storages
boto
boto3
botocore
django-filter
